+++ src/campaign/modulos/infraestructura/v1/comandos.py
@@ -46,3 +46,20 @@ class ComandoCancelarCampaign(ComandoIntegracion):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        
+# PARTNERS
+class CancelarPartner(Record):
+    id = String()
+
+class ComandoCancelarPartner(ComandoIntegracion):
+    id = String(default=str(uuid.uuid4()))
+    time = Long()
+    ingestion = Long(default=time_millis())
+    specversion = String(default="v1")
+    type = String(default="CancelarPartner")
+    datacontenttype = String()
+    service_name = String(default="partner.aeroalpes")
+    data = CancelarPartner
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
