+++ src/partner/modulos/infraestructura/dto.py
@@ -12,6 +12,7 @@ class Partner(Base):
     __tablename__ = "partners"
 
     id = Column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()), unique=True, index=True)
+    # id_campaign = Column(String(36), default=lambda: str(uuid.uuid4()))
     nombre = Column(String(100), nullable=False)
     tipo = Column(Enum(TipoPartner), nullable=False)
     informacion_perfil = Column(String(255))
